COM-3011: fix bugs
